Moved config in anticipation of tests
Moved in anticipation of functionality to check for and update broken/out of date config files.
